requirements.txt
mcp>=1.0.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
openai>=1.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
